requirements.txt
streamlit
google-generativeai
Pillow
pytesseract
GoogleNews
SpeechRecognition
deep-translator
firebase-admin
streamlit-oauth